Create ECRv2
There are two main functions that can be called:

depositEthscription(ethscriptionId: Bytes32): This function allows an account to deposit an ethscription.
transferEthscription(ethscriptionId: Bytes32, recipient: Address): This function allows the owner of an ethscription to transfer it to another account after a cooldown period.
These are the primary entry points for interacting with the contract. Users or other contracts can call these functions to deposit and transfer ethscriptions.

Additionally, there are events defined in the contract that emit information about important contract actions:

EthscriptionDeposited(owner: Address, ethscriptionId: Bytes32): This event is emitted when an ethscription is successfully deposited.
EthscriptionTransferred(previousOwner: Address, recipient: Address, ethscriptionId: Bytes32): This event is emitted when an ethscription is successfully transferred to a new owner.
These events can be monitored by external systems or dApps to track the state changes within the contract.